refactor(pr): tab 导航与各 tab 内容归拢到 pr/tabs/
features/pr/ 根目录混了工作区外壳、tab 导航、各 tab 内容、列表项、状态栏 chip。
把 tab 相关（PrTabs 导航 + diff / comments / drafts / CommitsPanel / PrInfoView
五个 tab 内容）收进 pr/tabs/ 子目录，根目录只剩工作区外壳（PrPanel / PrHeader /
PrEmpty / PrItem）+ statusbar/。纯目录迁移 + import 改写，无逻辑/界面变更。
