+++ apps/desktop/src/renderer/src/components/features/pr/tabs/diff/DiffView.tsx
@@ -1,6 +1,6 @@
 // 必须在用到 @monaco-editor/react 之前执行（loader.config 指向本地 monaco）。
 // 本文件经 React.lazy 动态加载，故 Monaco 随本 chunk 按需拉取，不进入口包。
-import '../../../../monaco-setup';
+import '../../../../../monaco-setup';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { createRoot, type Root } from 'react-dom/client';
@@ -21,23 +21,23 @@ import type {
   SyncProgressEvent,
 } from '@meebox/shared';
 import { policyForPlatform } from '@meebox/shared';
-import { invoke, subscribe } from '../../../../api';
-import { editorFontSize } from '../../../../editor-font';
-import { formatBackendError, type FormattedError } from '../../../../errors';
-import { REMOTE_REHYPE_PLUGINS } from '../../../../markdown';
-import { useDraftsForPr } from '../../../../stores/drafts-store';
-import { Avatar } from '../../../common/Avatar';
+import { invoke, subscribe } from '../../../../../api';
+import { editorFontSize } from '../../../../../editor-font';
+import { formatBackendError, type FormattedError } from '../../../../../errors';
+import { REMOTE_REHYPE_PLUGINS } from '../../../../../markdown';
+import { useDraftsForPr } from '../../../../../stores/drafts-store';
+import { Avatar } from '../../../../common/Avatar';
 import { DraftZone } from '../drafts/DraftZone';
-import { ErrorBoundary } from '../../../common/ErrorBoundary';
-import { makeBitbucketImageFor, transformBitbucketUrl } from '../../../common/BitbucketImage';
+import { ErrorBoundary } from '../../../../common/ErrorBoundary';
+import { makeBitbucketImageFor, transformBitbucketUrl } from '../../../../common/BitbucketImage';
 import { CommentEditEditor } from '../comments/CommentEditEditor';
 import { CommentReplyEditor } from '../comments/CommentReplyEditor';
-import { ConfirmModal } from '../../../common/ConfirmModal';
+import { ConfirmModal } from '../../../../common/ConfirmModal';
 import { DiffSearchPanel } from './DiffSearchPanel';
 import { FileTree } from './FileTree';
-import { PaneLoading } from '../../../common/Loading';
-import { FileTreeIcon, SearchIcon } from '../../../common/icons';
-import { languageFor } from '../../../../utils/language';
+import { PaneLoading } from '../../../../common/Loading';
+import { FileTreeIcon, SearchIcon } from '../../../../common/icons';
+import { languageFor } from '../../../../../utils/language';
 
 interface DiffViewProps {
   pr: StoredPullRequest;
